nexus: Sourcing Archaeological Materials by Chemical Composition

Exploration and analysis of compositional data in the framework of Aitchison (1986) <isbn: 9780412280603>. This package provides tools for chemical fingerprinting and source tracking of ancient materials.

Details

Package:	nexus

Type:	Package

Version:	0.0.0.9000

License:	GPL-3

Zenodo:

Package options

nexus uses the following options() to configure behavior:

nexus.autodetect: a logical scalar. Try to automatically assign values to the corresponding slot of a CompositionMatrix object when coercing a data.frame? Defaults to TRUE.

nexus.verbose: a logical scalar. Should R report extra information on progress? Defaults to TRUE.

See also

Author

Full list of authors and contributors (alphabetic order)

Nicolas Frerebeau	Université Bordeaux Montaigne, France

Brice Lebrun	Université Bordeaux Montaigne, France

Anne Philippe	Université de Nantes, France

Package maintainer

Nicolas Frerebeau
nicolas.frerebeau@u-bordeaux-montaigne.fr

Archéosciences Bordeaux (UMR 6034)
Maison de l'Archéologie
Université Bordeaux Montaigne
F-33607 Pessac cedex
France