Data Transformation

Data Transformation

transform_clr(object, ...)

transform_alr(object, ...)

transform_ilr(object, ...)

transform_pivot(object, ...)

# S4 method for CompositionMatrix
transform_clr(object, base = exp(1))

# S4 method for CompositionMatrix
transform_alr(object, j = ncol(object), base = exp(1))

# S4 method for CompositionMatrix
transform_ilr(object, base = exp(1))

# S4 method for CompositionMatrix
transform_pivot(object, pivot = 1, base = exp(1))

Arguments

object

A CompositionMatrix object.

...

Currently not used.

base

A positive numeric value giving the base with respect to which logarithms are computed. Change this only if you know what your are doing.

j

An integer giving the index of the rationing part.

pivot

An integer giving the index of the pivotal variable.

Value

A LogRatio object.

References

Aitchison, J. (1986). The Statistical Analysis of Compositional Data. London: Chapman and Hall. doi: 10.1007/978-94-009-4109-0 .

Egozcue, J. J., Pawlowsky-Glahn, V., Mateu-Figueras, G. & Barceló-Vidal, C. (2003). Isometric Logratio Transformations for Compositional Data Analysis. Mathematical Geology, 35(3), 279-300. doi: 10.1023/A:1023818214614 .

Fišerová, E. & Hron, K. (2011). On the Interpretation of Orthonormal Coordinates for Compositional Data. Mathematical Geosciences, 43(4), 455‑468. doi: 10.1007/s11004-011-9333-x .

Author

N. Frerebeau

Examples

## Coerce to chemical data
data("kommos", package = "folio")
coda <- as_composition(kommos[, -c(1, 2)])
coda <- remove_NA(coda, margin = 1)
#> 6 rows were removed:
#> * 98/42 (n=1)
#> * 98/45 (n=1)
#> * 98/49 (n=2)
#> * 98/50 (n=1)
#> * 98/71 (n=1)
#> * 98/81 (n=1)

## Centered log-ratio
clr <- transform_clr(coda)

## Additive log-ratio
alr <- transform_alr(coda, j = "Ca")

## Isometric log-ratio
ilr1 <- transform_ilr(coda)
ilr2 <- transform_pivot(coda)