Centered Log-Ratios (CLR)

Computes CLR transformation.

Usage

transform_clr(object, ...)

# S4 method for CompositionMatrix
transform_clr(object, weights = FALSE)

Arguments

object

A CompositionMatrix object.

...

Currently not used.

weights

A logical scalar: sould a varying weight be used. If FALSE (the default), equally-weighted parts are used. Alternatively, a positive numeric vector of weights can be specified.

Value

A CLR object.

Details

The CLR transformation computes the log of each part relative to the geometric mean of all parts.

References

Aitchison, J. (1986). The Statistical Analysis of Compositional Data. London: Chapman and Hall. doi:10.1007/978-94-009-4109-0 .

Greenacre, M. J. (2019). Compositional Data Analysis in Practice. Boca Raton: CRC Press.

Greenacre, M. J. (2021). Compositional Data Analysis. Annual Review of Statistics and Its Application, 8(1): 271-299. doi:10.1146/annurev-statistics-042720-124436 .

See also

Author

N. Frerebeau

Examples

## Coerce to compositional data
data("hongite")
coda <- as_composition(hongite)

## Pairwise log-ratio
lr <- transform_lr(coda)

## Centered log-ratio
clr <- transform_clr(coda)

## Additive log-ratio
alr <- transform_alr(coda)

## Isometric log-ratio
ilr <- transform_ilr(coda)
plr <- transform_plr(coda)

## Inverse transformation
inv_clr <- transform_inverse(clr)
all.equal(coda, inv_clr)
#> [1] TRUE

inv_alr <- transform_inverse(alr)
all.equal(coda, inv_alr)
#> [1] TRUE

inv_ilr <- transform_inverse(ilr)
all.equal(coda, inv_ilr)
#> [1] TRUE

inv_plr <- transform_inverse(plr)
all.equal(coda, inv_plr)
#> [1] TRUE