Coerce to a Closed Compositional Matrix

Coerces an object to a CompositionMatrix object.

Usage

as_composition(from, ...)

# S4 method for numeric
as_composition(from)

# S4 method for matrix
as_composition(from)

# S4 method for data.frame
as_composition(
  from,
  codes = NULL,
  samples = NULL,
  groups = NULL,
  auto = getOption("nexus.autodetect"),
  verbose = getOption("nexus.verbose")
)

Arguments

from

A matrix or data.frame to be coerced.

...

Currently not used.

codes

An integer giving the index of the column to be used as laboratory codes (unique identifiers).

samples

An integer giving the index of the column to be used for sample identification: allows duplicates to identify replicated measurements. If NULL (the default), row names will be used as sample IDs.

groups

An integer giving the index of the column to be used to group the samples. If NULL (the default), no grouping is stored.

auto

A logical scalar: try to automatically detect codes, samples and groups columns?

verbose

A logical scalar: should R report extra information on progress?

Value

A CompositionMatrix object.

Details

The CompositionMatrix class has special slots:

codes for laboratory codes,

samples for repeated measurements/observation,

groups to group data by site/area.

When coercing a data.frame to a CompositionMatrix object, an attempt is made to automatically assign values to these slots by mapping column names (case insensitive, plural insensitive). This behavior can be disabled by setting options(nexus.autodetect = FALSE) or overridden by explicitly specifying the columns to be used.

See vignette("nexus").

Note

All non-numeric variable will be removed.

See also

Author

N. Frerebeau

Examples

## Create a count matrix
A1 <- matrix(data = as.numeric(sample(1:100, 100, TRUE)), nrow = 20)

## Coerce to compositions
B <- as_composition(A1)

## Row sums are internally stored before coercing to relative frequencies
get_totals(B)
#>  [1] 262 217  56 363 232 198 276 264 244 312 286 204 171 270 234 173 204 317 252
#> [20] 229

## This allows to restore the source data
A2 <- as_amounts(B)

## Coerce to a data.frame
X <- data.frame(B)
head(X)
#>            V1        V2         V3         V4         V5
#> S1 0.12595420 0.3282443 0.17557252 0.25572519 0.11450382
#> S2 0.07834101 0.3364055 0.09677419 0.37788018 0.11059908
#> S3 0.14285714 0.1071429 0.07142857 0.51785714 0.16071429
#> S4 0.12947658 0.2644628 0.22589532 0.16804408 0.21212121
#> S5 0.06034483 0.3060345 0.28017241 0.02155172 0.33189655
#> S6 0.16161616 0.4646465 0.13131313 0.19696970 0.04545455