Row Sums

Retrieves or defines the row sums (before closure).

Usage

get_totals(x)

set_totals(x) <- value

# S4 method for CompositionMatrix
get_totals(x)

# S4 method for CompositionMatrix
set_totals(x) <- value

Arguments

x

An object from which to get or set totals.

value

A possible value for the totals of x.

Value

set_totals() returns an object of the same sort as x with the new row sums assigned.

get_totals() returns the row sums of x.

See also

Author

N. Frerebeau

Examples

## Create a count matrix
A1 <- matrix(data = as.numeric(sample(1:100, 100, TRUE)), nrow = 20)

## Coerce to compositions
B <- as_composition(A1)

## Row sums are internally stored before coercing to relative frequencies
get_totals(B)
#>  [1] 113 173 288 192 461 348 212 312 284 293 227 243 300 280 203 267 203 281 200
#> [20] 195

## This allows to restore the source data
A2 <- as_amounts(B)

## Coerce to a data.frame
X <- data.frame(B)
head(X)
#>            V1         V2         V3         V4        V5
#> S1 0.03539823 0.52212389 0.10619469 0.07964602 0.2566372
#> S2 0.22543353 0.06358382 0.02890173 0.17341040 0.5086705
#> S3 0.27430556 0.34722222 0.02083333 0.10763889 0.2500000
#> S4 0.46875000 0.12500000 0.17187500 0.12500000 0.1093750
#> S5 0.21691974 0.20390456 0.16919740 0.20173536 0.2082430
#> S6 0.14080460 0.18678161 0.27586207 0.28735632 0.1091954